refactored code: split the routes to match modles and controllers structure, moved jwt and credentials validation to middleware, fixed the transactions POST items to received params and switched to GET requests, and updated documentation
