+++ models/usermodels.js
@@ -46,24 +46,6 @@ function findUser(req,res,next) {
         })
 }
 
-function validateCredentials (req, res, next) {
-    const {id, email} = req.user
-    knex('users_list')
-        .where({email: email})
-        .then((userInfo) => {
-            if (userInfo.length === 0) {
-                return res.status(400).send("There is no user registered with this email")
-            }
-            if (id !== userInfo[0].id) {
-                return res.status(400).send('Invalid token')
-            }
-            next()
-        })
-        .catch((err) => {
-            res.status(400).send("Something wrong went with your request.")
-        })
-}
-
 function findBankList (req,res) {
     const {id, email} = req.user
     knex('opening_bank_balances')
@@ -90,35 +72,11 @@ function addBankAcc (req,res) {
     })
 }
 
-function findBankAcc(req,res,next) {
-    const {id, email} = req.user
-    const {debit, credit, bank_type} = req.body
-    
-    let searchAcc = ""
-    if (bank_type === "c") { searchAcc = credit}
-    if (bank_type === "d") { searchAcc = debit}
-
-    knex('opening_bank_balances')
-        .where({user_id:id, acc_des: searchAcc})
-        .then( (userInfo) => {
-            if (userInfo.length === 0) {
-                return res.status(400).send("The specified account does not exist for the user mentioned")
-            }
-
-            next();
-        })
-        .catch((err) => {
-            res.status(400).send("We ran into difficulties searching for account info")
-        })
-}
-
 module.exports = {
     readFiles,
     writeFiles,
     addNewUser,
     findUser,
-    validateCredentials,
     addBankAcc,
-    findBankAcc,
     findBankList
 }
