--- routes/routes.js
@@ -1,67 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.use(express.static('public'));
-
-const {sendAPIDoc, sendAccountList} = require('../controllers/general')
-
-const {userRegistration, userLogin, authenticateUser, addBankInfo} = require('../controllers/usersignin');
-const {addNewUser, findUser, validateCredentials, addBankAcc, findBankAcc, findBankList} = require('../models/usermodels');
-
-const decodeJWT = require('../controllers/decodeJWT');
-
-const {checkTransactions, addTransactions, checkTranPeriod, sendTotalByPeriod} = require('../controllers/transactions')
-const {addNewTran, findTranByPeriod, findDebitByPeriod, findCreditByPeriod} = require('../models/transactionsmodels')
-
-
-router
-    .get('/', sendAPIDoc)
-    .get('/account-list', sendAccountList)
-    .post('/user/register', 
-        userRegistration, 
-        addNewUser
-    )
-    //get
-    .post('/user/login', 
-        userLogin, 
-        findUser, 
-        authenticateUser
-    )
-    .post('/user/initial-set-up', 
-        decodeJWT, 
-        validateCredentials, 
-        addBankInfo, 
-        addBankAcc
-    )
-    .get('/user/bank-list',
-        decodeJWT,
-        validateCredentials,
-        findBankList
-    )
-    .post('/user/transaction', 
-        decodeJWT, 
-        validateCredentials, 
-        checkTransactions, 
-        findBankAcc, 
-        addTransactions,
-        addNewTran
-    )
-    //get
-    .post('/user/transaction-by-period',
-        decodeJWT,
-        validateCredentials,
-        checkTranPeriod,
-        findTranByPeriod
-    )
-    //get
-    .post('/user/total-by-period',
-        decodeJWT,
-        validateCredentials,
-        checkTranPeriod,
-        findDebitByPeriod,
-        findCreditByPeriod,
-        sendTotalByPeriod
-    )
-    //.get('/user/initial-set-up', decodeJWT, addBankInfo)
-
-module.exports = router;
